refactor(ui): commonize media sidebar (#298)
